+++ changelog.md
@@ -1,3 +1,4 @@
+## 1.1.0
 - Improve support for multiple lines content
   - **Breaking Change** User can not directly change content via `tag.contnt = "abc"`
     - As content can be multiple lines, they are now a part of elements' sons